refact: buildMainTreeルーチンのリファクタリング
